backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-cors-headers==4.0.0
django-environ==0.10.0
django-import-export==3.2.0
django-parler==2.3
django-parler-rest==2.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
mailjet==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
typing_extensions==4.6.3
xlrd==2.0.1
xlwt==1.3.0
